+++ site/dat/docs/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.16.25<sup> 22 August 2023</sup>
+- Latest chromedriver support
+- Cloud test csv report fix
+
 ## 1.16.24<sup> 8 August 2023</sup>
 - Selenium 4.10.0 support
 - Upgrade to ruby 3.2.2